+++ mindcode/src/main/java/info/teksol/mindcode/compiler/Remarks.java
@@ -1,0 +1,7 @@
+package info.teksol.mindcode.compiler;
+
+public enum Remarks {
+    NONE,
+    PASSIVE,
+    ACTIVE
+}