+++ mindcode/src/test/java/info/teksol/mindcode/compiler/AbstractGeneratorTest.java
@@ -209,7 +209,8 @@ public class AbstractGeneratorTest extends AbstractAstTest {
     protected final AstContext mockAstRootContext = AstContext.createRootNode();
     protected final AstContext mockAstContext = mockAstRootContext.createSubcontext(AstSubcontextType.BASIC, 1.0);
 
-    protected final InstructionProcessor ip = createInstructionProcessor(createCompilerProfile(), s ->{});
+    protected final CompilerProfile profile = createCompilerProfile();
+    protected final InstructionProcessor ip = createInstructionProcessor(profile, s ->{});
 
     protected final LogicInstruction createInstruction(Opcode opcode) {
         return ip.createInstruction(mockAstContext, opcode);